feature: adding user validation to delete guide

+++ usecases/src/main/kotlin/com/organizee/usecases/guide/impl/RemoveGuideUsecaseImpl.kt
@@ -3,6 +3,7 @@ package com.organizee.usecases.guide.impl
 import com.organizee.boundaries.db.services.GuideService
 import com.organizee.boundaries.search.SearchService
 import com.organizee.usecases.guide.RemoveGuideUseCase
+import com.organizee.usecases.guide.commands.DeleteGuideCommand
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
@@ -17,10 +18,14 @@ class RemoveGuideUsecaseImpl(
         val logger: Logger = LoggerFactory.getLogger(this::class.java)
     }
 
-    override fun execute(input: String) {
+    override fun execute(input: DeleteGuideCommand) {
         logger.info("Removing guide | slug=$input")
-        val guide = guideService.getGuide(input)
+        val guide = guideService.getGuide(input.slug)
+
+        if (guide.user?.id != input.userId)
+            throw Exception("Guide is not owned by user")
+
         searchService.delete(guide)
-        guideService.removeGuide(input)
+        guideService.removeGuide(input.slug)
     }
 }
